refactor(editor): rename generateFn to fillFunction and make it more flexible
