+++ app/lib/app/SecvisogramPage/View/FormEditor/editors/GenericEditor.js
@@ -14,7 +14,7 @@ import CVSS2Editor from './GenericEditor/CVSS2Editor.js'
 import CVSSV3Attribute from './GenericEditor/Attributes/CVSS3Attribute.js'
 import AppConfigContext from '../../../../shared/context/AppConfigContext.js'
 import UserInfoContext from '../../../../shared/context/UserInfoContext.js'
-import { uniqueProductId } from '../shared/unique-id.js'
+import { uniqueProductId } from '../shared/fillFieldFunctions.js'
 
 /**
  * utility function to get the color of circles identifying errors
@@ -51,7 +51,7 @@ export default function Editor({
   const { loginAvailable } = React.useContext(AppConfigContext)
   const userInfo = React.useContext(UserInfoContext)
 
-  const { doc, collectIds } = React.useContext(DocumentEditorContext)
+  const { doc, updateDoc, collectIds } = React.useContext(DocumentEditorContext)
 
   const uiType = property.metaData?.uiType
   const enableLast = uiType === 'ARRAY_REVISION_HISTORY'
@@ -203,7 +203,7 @@ export default function Editor({
           value={value || ''}
           property={property}
           disabled={disabled}
-          generateFn={uniqueProductId}
+          fillFunction={() => updateDoc(instancePath, uniqueProductId())}
         />
       ))
     } else if (uiType === 'STRING_PRODUCT_ID') {
